refactoring defaults
