+++ magpylib/_src/defaults/__init__.py
@@ -1,0 +1,1 @@
+"""defaults module"""
