+++ magpylib/_src/defaults/defaults_classes.py
@@ -1,4 +1,4 @@
-from magpylib._src.default_utils import (
+from magpylib._src.defaults.defaults_utility import (
     MagicProperties,
     validate_property_class,
     color_validator,
