+++ magpylib/_src/defaults/defaults_utility.py
@@ -2,7 +2,7 @@
 
 from copy import deepcopy
 import collections.abc
-from magpylib._src.defaults import DEFAULTS
+from magpylib._src.defaults.defaults_values import DEFAULTS
 
 
 SUPPORTED_PLOTTING_BACKENDS = ("matplotlib", "plotly")
