+++ magpylib/_src/defaults/defaults_values.py
