+++ magpylib/_src/display/display.py
@@ -4,7 +4,7 @@ import warnings
 from magpylib._src.utility import format_obj_input, test_path_format
 from magpylib._src.display.display_matplotlib import display_matplotlib
 from magpylib._src.input_checks import check_dimensions
-from magpylib._src.default_classes import default_settings as Config
+from magpylib._src.defaults.defaults_classes import default_settings as Config
 
 
 # ON INTERFACE
