+++ magpylib/_src/display/display_plotly.py
@@ -18,7 +18,7 @@ except ImportError as missing_module:  # pragma: no cover
 import numpy as np
 from scipy.spatial.transform import Rotation as RotScipy
 from magpylib import _src
-from magpylib._src.default_classes import default_settings as Config
+from magpylib._src.defaults.defaults_classes import default_settings as Config
 from magpylib._src.display.plotly_sensor_mesh import get_sensor_mesh
 from magpylib._src.style import (
     get_style,
@@ -32,7 +32,7 @@ from magpylib._src.display.display_utility import (
     draw_arrowed_circle,
     place_and_orient_model3d,
 )
-from magpylib._src.default_utils import (
+from magpylib._src.defaults.defaults_utility import (
     SIZE_FACTORS_MATPLOTLIB_TO_PLOTLY,
     linearize_dict,
 )
