+++ magpylib/_src/obj_classes/class_BaseTransform.py
@@ -5,7 +5,7 @@
 import numpy as np
 from scipy.spatial.transform import Rotation as R
 from magpylib._src.exceptions import MagpylibBadUserInput
-from magpylib._src.default_classes import default_settings as Config
+from magpylib._src.defaults.defaults_classes import default_settings as Config
 from magpylib._src.input_checks import (
     check_start_type,
     check_increment_type,
