+++ magpylib/_src/obj_classes/class_Collection.py
@@ -9,7 +9,7 @@ from magpylib._src.utility import (
 from magpylib._src.obj_classes.class_BaseGeo import BaseGeo
 from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
 from magpylib._src.fields.field_wrap_BH_level2 import getBH_level2
-from magpylib._src.default_utils import validate_style_keys
+from magpylib._src.defaults.defaults_utility import validate_style_keys
 from magpylib._src.exceptions import MagpylibBadUserInput
